+++ lib/brazil.dart
@@ -2,3 +2,4 @@
 library brazil;
 
 export 'src/brazil/cep_field.dart';
+export 'src/brazil/cpf_field.dart';