+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Dart Package Versioning](https://dart.dev/tools/pub
 
 ## [Unreleased]
 
+### Added
+
+- CpfField, a text form field for entering CPF values —
+  [31](https://github.com/dartoos-dev/well_formed/issues/31)
+
 ## [0.3.2] - 2021-08-13
 
 ### Added